+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Upcoming
+* Added block data
+  * In the `data/block_data` folder can now be placed JSON to modify blocks properties, such as hardness, explosion resistance, speed and jump factor and a few more
+  * With this, harder crops, custom hardness and terrain slowdown has been removed and moved to a brand new data pack (block_data)
+    * To make crops able to be broken faster by hoes (and not axes) they still must be added in the `iguanatweaksreborn:harder_crops` block tag
 * Sponges
   * Wet Sponges now dry when exposed to sun and Sponges now become wet when exposed to rain 
 * Graves